Deployment manifests for argo-workflows
This commit adds the manifests to deploy argo-workflows in OKD and adds integrates the OKD's oauth-proxy as a reverseproxy to execute a simple authentication against OKD before accessing the argo workflows UI and services. Although it works, this should still considered an experimental work not to be used in production clusters. This commit needs refactoring and improvements.